Add bloom and lighting
Sun is the origin of light in the scene. This is done via point light. There were some major problems with getting the sun to light up, since the point light was inside the sun. As a workaround, bloom was added and a lot of emission to the suns model was added so that the sun glows bright. Downside of this is that the suns texture cannot be seen anymore.